Fix double-counting of changes
The code to update each edge in a single commit was updating (non-self)
edges twice because of a flaw in the way the "iteration" was being done.
It was based on some earlier code that used a matrix, and that produced
duplicate, reversed edges. When this code was replaced, the 2D
iteration should have been updated so as to exclude the duplicate edges.
This fault was hidden by the fact that the edgeFor() function was
finding the same link for both cases (both directions), so it appeared
to be working. But the weight in the edge was still being updated in
both cases, which produced double counting.
The perils of trying to hack the code rather than TDD it?