Graph analysis excludes self-links
When analysing the edges to see which files have changed most with
other files, don't include edges that link from and to the same node.
These are not relevant, they just track any stats of files that are
changed, not how they were changed in relation to other files.
Separately, we currently also exclude them from the displayed graph.