Short filename should be a display choice
Node shouldn't store the short filename - that is just the filename,
without the path, it should store the full path/name.
We want to distinguish between different files (in different paths)
with the same name.
Displaying just the short name on the visualisation is a display
choice, not part of the internals of what we store.